services/tts/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
coqui-tts>=0.24.0
piper-tts>=1.2.0
numpy>=1.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
httpx>=0.25.0
